Added an external library to mock out Firebase. A few problems with it: - it does not have all of the same methods as the real one,   - the API uses a set function that works to both create and update documents. The function that's getting called in the mock libary *only* updates, so there are errors there   - not as important, it lacks the collection_group function that is used to look across collections in multiple parents
The solution to these problems is probably just to patch the code ourselves. Which would be fine for a longer-term project.

Because of the singleton design of the api class, using a mock of its database sets that as the database for *all tests*, so a teardown method would be required to recreate the api between tests (or modules of tests if the whole module uses the mock). The fixture could be modified to take in a boolean indicating if it should be the real or the mock that's returned by the method.